NOVA_BACKEND/pyproject.toml
[project]
name = "nova"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "boto3>=1.40.53",
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "openai>=2.29.0",
    "psycopg>=3.3.3",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "isort>=8.0.1",
    "ruff>=0.15.5",
]
